content/pyproject.toml
[tool.poetry]
name = "md-frontmatter-updater"
version = "0.1.0"
description = "A script to add frontmatter to markdown files"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
